requirements.txt
# Web框架
Flask==2.0.1
Werkzeug==2.0.1
Flask-Session==0.4.0
Flask-Cors==3.0.10

# 数据库
mysql-connector-python==8.0.26
PyMySQL==1.0.2

# 人脸识别和图像处理
face-recognition==1.3.0
face-recognition-models==0.3.0
dlib>=19.22.0
opencv-python==4.5.3.56
numpy==1.21.2
Pillow==8.3.2

# 深度学习模型
deepface==0.0.79
tensorflow>=2.6.0
keras>=2.6.0

# 数据处理和分析
pandas>=1.3.0
scipy>=1.7.0

# 环境变量和配置
python-dotenv==0.19.0
PyYAML>=5.4.1

# 工具包
requests>=2.26.0
tqdm>=4.62.0
typing-extensions>=3.10.0

# 开发工具
pytest==7.0.1
black==22.3.0
flake8==4.0.1
pylint==2.9.6

# 文档
Sphinx==4.2.0